Re-added camos and fixed the Fang
The Commando and the Riot shields now have the camos that the OPTRE parents do.
Etilka now fires ball instead of explosive.
M6C and M6G laser attachments are fixed to fully swap between laser, IR, and flashlight.
Fixed the SRS Fang's inheritance isses, removed the 19th's scope from it due to lack of model and the OPTRE scopes from the SRSAM